move out more fns
